v2.1.5.4
NMS.getAbsorptionHearts, NMS.setAbsorptionHearts 추가
플렉터 능력 및 설명 추가
구속 설명 추가
관통화살, 스나이퍼의 대미지에 힘 마법 부여가 적용되도록 업데이트
서버 리로드시 발생하는 콘피그 관련 오류 해결
Deflectable.getShooter 추가
Messager 유틸 사용시 새로운 객체를 만들지 않도록 변경
EnchantLib.getDamageWithPowerEnchantment 추가
Locationutil.collectEntities 최적화

+++ modules/Plugin/src/daybreak/abilitywar/ability/list/ShowmanShip.java
@@ -49,7 +49,7 @@ public class ShowmanShip extends AbilityBase {
 	private static final Type[] types = {
 			Type.BALL_LARGE, Type.STAR
 	};
-	private static final double radius = 7;
+	private static final int radius = 7;
 	private final RGB WEAK = new RGB(214, 255, 212);
 	private final RGB POWER = new RGB(255, 184, 150);
 	private final RGB POWERFUL = new RGB(255, 59, 59);
@@ -120,7 +120,7 @@ public class ShowmanShip extends AbilityBase {
 
 	private final Predicate<Entity> unequalPredicate = Predicates.PARTICIPANTS_UNEQUAL(getPlayer());
 
-	private double getPoint(double horizontal, double vertical) {
+	private double getPoint(int horizontal, int vertical) {
 		Location center = getPlayer().getLocation();
 		double centerX = center.getX(), centerZ = center.getZ();
 		double point = 0;
